カウントダウン表示の関数を分離
